Update after testing
It turns out that the ET Systems 3 Channel received only has normally open
contacts on the relays. So this has been compensated for in the code by inverting the values that we are expecting.

The config file has been updated with comments showing which arduino pins each
sensor/input is assigned to.